+++ integers/adder_test.go
@@ -1,6 +1,9 @@
 package integers
 
-import "testing"
+import (
+	"fmt"
+	"testing"
+)
 
 func TestAdder(t *testing.T) {
 	sum := Add(2, 2)
@@ -8,4 +11,10 @@ func TestAdder(t *testing.T) {
 	if sum != expected {
 		t.Errorf("expected '%d' but got '%d'", expected, sum)
 	}
+}
+
+func ExampleAdd() {
+	sum := Add(1, 5)
+	fmt.Println(sum)
+	// Output: 6
 }
