+++ iteration/repeat_test.go
@@ -1,6 +1,7 @@
 package iteration
 
 import (
+	"fmt"
 	"testing"
 )
 
@@ -10,5 +11,16 @@ func TestRepeat(t *testing.T) {
 
 	if repeat != expected {
 		t.Errorf("expected %q but got %q", expected, repeat)
+	}
+}
+func ExampleRepeat() {
+	repeat := Repeat("a", 9)
+	fmt.Printf(repeat)
+	// Output: aaaaaaaaa
+}
+
+func BenchmarkRepeat(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		Repeat("a", 8)
 	}
 }
